// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ResourceType = {
  DATA_ROOM: 'DATA_ROOM',
  FOLDER: 'FOLDER',
  FILE: 'FILE'
} as const

export type ResourceType = (typeof ResourceType)[keyof typeof ResourceType]


export const ShareRole = {
  VIEWER: 'VIEWER'
} as const

export type ShareRole = (typeof ShareRole)[keyof typeof ShareRole]